Move index.ts to compiler and then introduce a real index
