+++ src/__tests__/compiler.test.ts
@@ -1,9 +1,9 @@
 import { describe, it } from "node:test";
 import { e2eVoidText } from "./fixtures/e2e-file.js";
-import { compileText } from "../index.js";
+import { compileText } from "../compiler.js";
 import { strict as assert } from "node:assert";
 
-describe("API: E2E", () => {
+describe("Compiler: E2E", () => {
   it("Can compile and run a basic void file", async () => {
     const mod = await compileText(e2eVoidText);
     const bin = mod.emitBinary();
